# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EasyAudit-Next API",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {"operationId": "login", "tags": ["auth"], "responses": ["200", "401", "429"]}
    },
...
  }
}